pyproject.toml
[project]
name = "renpak"
version = "0.1.0"
description = "Ren'Py game asset compression toolkit — JPG/PNG → AVIF transcoding"
requires-python = ">=3.11"
license = "MPL-2.0"
dependencies = [
    "pillow>=10.0",
    "pillow-avif-plugin>=1.4",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[project.scripts]
renpak = "renpak.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["python/renpak"]
